refactor: 로또 당첨 번호 추첨을 LotteryMachine으로 관리
